Convert all Player ObjectID's to units, add 0xE0000000 check

+++ NoTankYou/Utilities/PlayerLocator.cs
@@ -4,9 +4,9 @@ namespace NoTankYou.Utilities
 {
     internal static class PlayerLocator
     {
-        public static PlayerCharacter? GetPlayer(int objectId)
+        public static PlayerCharacter? GetPlayer(uint objectId)
         {
-            var result = Service.ObjectTable.SearchById((uint) objectId);
+            var result = Service.ObjectTable.SearchById(objectId);
 
             if (result?.GetType() == typeof(PlayerCharacter))
                 return result as PlayerCharacter;
